make moduler base the script creat two files form.js and todos.js
