+++ src/style.css
@@ -13,7 +13,7 @@
 
 #todo-list {
   padding-left: 1px !important;
-  margin: 5px 1px;
+  margin: 6px 1px;
 }
 
 .newtodo {
@@ -109,23 +109,23 @@
   align-items: center;
   align-self: center;
   width: 96%;
-  padding: 5px;
-  margin: 10px;
+  padding: 10px;
+  margin: 15px 5px;
 }
 
 .myCheckbox, .description {
   display:  inline-block;
 }
 
-#editTodoInput {
+#errorMessag {
+  color: red;
+}
+
+.editInput {
   padding: 5px;
   border: none;
-  width: 80%;
+  width: 78%;
   border: none;
   outline: none;
   border-bottom: 1px solid gray;
-}
-
-#errorMessag {
-  color: red;
 }
